Need help with repainting...
So I'm to the point where I can start working on saving and opening, but
I cannot, for some reason, refresh the JList in my program. It refuses
to. I tried repainting my listScroller too, but neither decided to
update. I don't know if I explained to you what my intention of this
program was, soooo her it goes. I am completely obsessed with Netflix
and Amazon Video so my idea was to create a tracker for your shows which
will allow you to set the name, a description, a picture for the show,
and how many seasons there are for the certain show. Now I beg for your
dearest assistance and hope you have fun today and tomorrow! :)